# packages/contracts/openapi/openapi.json
{"components": {"schemas": {"Employee": {"properties": {"binding": {"$ref": "#/components/schemas/EmployeeBinding"}, "created_at": {"format": "date-time", "title": "Created At", "type": "string"}, "current_revision_id": {"title": "Current Revision Id", "type": "string"}, "display_name": {"title": "Display Name", "type": "string"}, "grants": {"items": {"$ref": "#/components/schemas/Grant"}, "title": "Grants", "type": "array"}, "id": {"title": "Id", "type": "string"}, "revision": {"$ref": "#/components/schemas/EmployeeRevision"}, "status": {"enum": ["active", "paused", "archived"], "title": "Status", "type": "string"}, "workspace_id": {"title": "Workspace Id", "type": "string"}}, "required": ["id", "workspace_id", "display_name", "status", "current_revision_id", "created_at", "revision", "binding", "grants"], "title": "Employee", "type": "object"}, "EmployeeBinding": {"properties": {"created_at": {"format": "date-time", "title": "Created At", "type": "string"}, "dsh_agent_id": {"title": "Dsh Agent Id", "type": "string"}, "dsh_session_id": {"title": "Dsh Session Id", "type": "string"}, "employee_id": {"title": "Employee Id", "type": "string"}, "id": {"title": "Id", "type": "string"}, "memory_scope_id": {"title": "Memory Scope Id", "type": "string"}}, "required": ["id", "employee_id", "dsh_agent_id", "dsh_session_id", "memory_scope_id", "created_at"], "title": "EmployeeBinding", "type": "object"}, "EmployeeCreate": {"properties": {"display_name": {"maxLength": 120, "minLength": 1, "title": "Display Name", "type": "string"}, "grants": {"items": {"$ref": "#/components/schemas/GrantCreate"}, "title": "Grants", "type": "array"}, "model": {"maxLength": 200, "minLength": 1, "title": "Model", "type": "string"}, "responsibility": {"maxLength": 4000, "minLength": 1, "title": "Responsibility", "type": "string"}, "runtime_profile": {"enum": ["workspace_read", "workspace_write", "network_denied"], "title": "Runtime Profile", "type": "string"}}, "required": ["display_name", "responsibility", "runtime_profile", "model"], "title": "EmployeeCreate", "type": "object"}, "EmployeeRevise": {"properties": {"grants": {"items": {"$ref": "#/components/schemas/GrantCreate"}, "title": "Grants", "type": "array"}, "model": {"maxLength": 200, "minLength": 1, "title": "Model", "type": "string"}, "responsibility": {"maxLength": 4000, "minLength": 1, "title": "Responsibility", "type": "string"}, "runtime_profile": {"enum": ["workspace_read", "workspace_write", "network_denied"], "title": "Runtime Profile", "type": "string"}}, "required": ["responsibility", "runtime_profile", "model"], "title": "EmployeeRevise", "type": "object"}, "EmployeeRevision": {"properties": {"created_at": {"format": "date-time", "title": "Created At", "type": "string"}, "employee_id": {"title": "Employee Id", "type": "string"}, "id": {"title": "Id", "type": "string"}, "model": {"title": "Model", "type": "string"}, "responsibility": {"title": "Responsibility", "type": "string"}, "revision_number": {"title": "Revision Number", "type": "integer"}, "runtime_profile": {"title": "Runtime Profile", "type": "string"}}, "required": ["id", "employee_id", "revision_number", "responsibility", "runtime_profile", "model", "created_at"], "title": "EmployeeRevision", "type": "object"}, "ErrorDetail": {"properties": {"code": {"title": "Code", "type": "string"}, "correlation_id": {"title": "Correlation Id", "type": "string"}, "message": {"title": "Message", "type": "string"}}, "required": ["code", "message", "correlation_id"], "title": "ErrorDetail", "type": "object"}, "ErrorEnvelope": {"properties": {"error": {"$ref": "#/components/schemas/ErrorDetail"}}, "required": ["error"], "title": "ErrorEnvelope", "type": "object"}, "Grant": {"properties": {"action": {"title": "Action", "type": "string"}, "employee_revision_id": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Employee Revision Id"}, "id": {"title": "Id", "type": "string"}, "level": {"enum": [0, 1, 2, 3], "title": "Level", "type": "integer"}, "requires_approval": {"title": "Requires Approval", "type": "boolean"}, "resource_kind": {"title": "Resource Kind", "type": "string"}, "resource_values": {"items": {"type": "string"}, "title": "Resource Values", "type": "array"}}, "required": ["id", "employee_revision_id", "action", "level", "resource_kind", "resource_values", "requires_approval"], "title": "Grant", "type": "object"}, "GrantCreate": {"properties": {"action": {"maxLength": 120, "minLength": 1, "title": "Action", "type": "string"}, "level": {"enum": [0, 1, 2, 3], "title": "Level", "type": "integer"}, "requires_approval": {"title": "Requires Approval", "type": "boolean"}, "resource_kind": {"minLength": 1, "title": "Resource Kind", "type": "string"}, "resource_values": {"items": {"type": "string"}, "title": "Resource Values", "type": "array"}}, "required": ["action", "level", "resource_kind", "resource_values", "requires_approval"], "title": "GrantCreate", "type": "object"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "title": "Detail", "type": "array"}}, "title": "HTTPValidationError", "type": "object"}, "HealthResponse": {"properties": {"service": {"const": "dsh-company", "title": "Service", "type": "string"}, "status": {"const": "ok", "title": "Status", "type": "string"}}, "required": ["status", "service"], "title": "HealthResponse", "type": "object"}, "ValidationError": {"properties": {"ctx": {"title": "Context", "type": "object"}, "input": {"title": "Input"}, "loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "title": "Location", "type": "array"}, "msg": {"title": "Message", "type": "string"}, "type": {"title": "Error Type", "type": "string"}}, "required": ["loc", "msg", "type"], "title": "ValidationError", "type": "object"}, "Workspace": {"properties": {"created_at": {"format": "date-time", "title": "Created At", "type": "string"}, "id": {"title": "Id", "type": "string"}, "name": {"title": "Name", "type": "string"}}, "required": ["id", "name", "created_at"], "title": "Workspace", "type": "object"}, "WorkspaceCreate": {"properties": {"name": {"maxLength": 120, "minLength": 1, "title": "Name", "type": "string"}}, "required": ["name"], "title": "WorkspaceCreate", "type": "object"}}}, "info": {"title": "DSH Company Service", "version": "0.1.0"}, "openapi": "3.1.0", "paths": {"/employees/{employee_id}": {"get": {"operationId": "get_employee_employees__employee_id__get", "parameters": [{"in": "path", "name": "employee_id", "required": true, "schema": {"title": "Employee Id", "type": "string"}}], "responses": {"200": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Employee"}}}, "description": "Successful Response"}, "404": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}, "description": "Not Found"}, "422": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}, "description": "Validation Error"}}, "summary": "Get Employee", "tags": ["company"]}}, "/employees/{employee_id}/revisions": {"post": {"operationId": "revise_employee_employees__employee_id__revisions_post", "parameters": [{"in": "path", "name": "employee_id", "required": true, "schema": {"title": "Employee Id", "type": "string"}}], "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/EmployeeRevise"}}}, "required": true}, "responses": {"200": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Employee"}}}, "description": "Successful Response"}, "404": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}, "description": "Not Found"}, "422": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}, "description": "Validation Error"}}, "summary": "Revise Employee", "tags": ["company"]}}, "/health": {"get": {"operationId": "health_health_get", "responses": {"200": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/HealthResponse"}}}, "description": "Successful Response"}}, "summary": "Health", "tags": ["foundation"]}}, "/workspaces": {"get": {"operationId": "list_workspaces_workspaces_get", "responses": {"200": {"content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/Workspace"}, "title": "Response List Workspaces Workspaces Get", "type": "array"}}}, "description": "Successful Response"}}, "summary": "List Workspaces", "tags": ["company"]}, "post": {"operationId": "create_workspace_workspaces_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/WorkspaceCreate"}}}, "required": true}, "responses": {"201": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Workspace"}}}, "description": "Successful Response"}, "422": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}, "description": "Validation Error"}}, "summary": "Create Workspace", "tags": ["company"]}}, "/workspaces/{workspace_id}": {"get": {"operationId": "get_workspace_workspaces__workspace_id__get", "parameters": [{"in": "path", "name": "workspace_id", "required": true, "schema": {"title": "Workspace Id", "type": "string"}}], "responses": {"200": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Workspace"}}}, "description": "Successful Response"}, "404": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}, "description": "Not Found"}, "422": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}, "description": "Validation Error"}}, "summary": "Get Workspace", "tags": ["company"]}}, "/workspaces/{workspace_id}/employees": {"get": {"operationId": "list_employees_workspaces__workspace_id__employees_get", "parameters": [{"in": "path", "name": "workspace_id", "required": true, "schema": {"title": "Workspace Id", "type": "string"}}], "responses": {"200": {"content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/Employee"}, "title": "Response List Employees Workspaces  Workspace Id  Employees Get", "type": "array"}}}, "description": "Successful Response"}, "404": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}, "description": "Not Found"}, "422": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}, "description": "Validation Error"}}, "summary": "List Employees", "tags": ["company"]}, "post": {"operationId": "create_employee_workspaces__workspace_id__employees_post", "parameters": [{"in": "path", "name": "workspace_id", "required": true, "schema": {"title": "Workspace Id", "type": "string"}}], "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/EmployeeCreate"}}}, "required": true}, "responses": {"201": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Employee"}}}, "description": "Successful Response"}, "404": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}, "description": "Not Found"}, "422": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}, "description": "Validation Error"}}, "summary": "Create Employee", "tags": ["company"]}}}}